승경 코드 병합

+++ PickMeal/src/main/java/PickMeal/PickMeal/controller/CommentController.java
@@ -3,7 +3,9 @@ package PickMeal.PickMeal.controller;
 import PickMeal.PickMeal.domain.Comment;
 import PickMeal.PickMeal.domain.User;
 import PickMeal.PickMeal.service.CommentService;
+import PickMeal.PickMeal.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,22 +17,30 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequiredArgsConstructor
 public class CommentController {
     private final CommentService commentService;
+    private final UserService userService;
 
     @PostMapping("/write/{boardId}")
-    public String writeComment(@PathVariable long boardId, Comment comment, @AuthenticationPrincipal User user) {
-        comment.setBoardId(boardId);
-        comment.setUser_id(user.getUser_id());
-        commentService.writeComment(comment);
+    public String writeComment(@PathVariable long boardId, Comment comment, Authentication authentication) {
+        User user = userService.getAuthenticatedUser(authentication);
+
+        if (user != null) {
+            comment.setBoardId(boardId);
+            comment.setUser_id(user.getUser_id());
+            commentService.writeComment(comment);
+        }
         return "redirect:/board/detail/" + boardId;
     }
 
     @PostMapping("/delete/{comment_id}")
-    public String deleteComment(@PathVariable long comment_id) {
-        long boardId = commentService.getCommentByComment_id(comment_id).getBoardId();
+    public String deleteComment(@PathVariable long comment_id, Authentication authentication) {
+        Comment comment = commentService.getCommentByComment_id(comment_id);
+        User user = userService.getAuthenticatedUser(authentication);
 
-        commentService.deleteComment(comment_id);
-
-        return "redirect:/board/detail/" + boardId;
+        // 로그인한 유저와 댓글 작성자가 같은지 확인하는 로직이 필요합니다!
+        if (user != null && comment.getUser_id() == user.getUser_id()) {
+            commentService.deleteComment(comment_id);
+        }
+        return "redirect:/board/detail/" + comment.getBoardId();
     }
